Separate system plugin and system status mappers
